Added Rule Book Folder
I added the rule book folder and within it added a few files. The first is the rule guide which will contain a majority of the rules and how to play. The second is the combat rule book and will contain the rest of the rules about combat. Th reason for the multiple files is because the rules of combat will change a lot